\name{EasyMARK-package}
\alias{EasyMARK-package}
\alias{EasyMARK}
\docType{package}
\title{
Make it easier to create, work with, and analyse capture histories,
with bias towards natural selection studies. 
}
\description{
Contains a few utility functions for working with capture-history data, 
a function for simulating capture-history data, and a function to fit this data, 
using a Gibbs sampler. 
}
\details{
\tabular{ll}{
Version: \tab 1.0\cr
Imports: MASS, stringr, rjags, coda, foreach, doParallel, random
Package: \tab EasyMARK\cr
Type: \tab Package\cr
Date: \tab 2014-01-29\cr
License: \tab GPL-2\cr
}

}
\author{
Maintainer: <john.waller@biol.lu.se>
}
\keyword{ package }

